+++ packages/tts-editor/CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `require` can resolve `index.lua` files without explicitly including them in the module name.
+- Command "Go to Last Error" added which opens the error line of the last reported error message.
 
 ## [1.1.0] - 2024-01-16